+++ src/app/controllers/index.ts
@@ -1,4 +1,6 @@
+export * from './dto.controller';
 export * from './feedback.controller';
 export * from './file.controller';
 export * from './list-files.controller';
+export * from './orm.controller';
 export * from './terminal.controller';