+++ src/app/providers/index.ts
@@ -1,4 +1,5 @@
 export * from './feedback.provider';
+export * from './list-dtos.providers';
 export * from './list-entities.providers';
 export * from './list-files.providers';
 export * from './list-methods.providers';